refactor(desktop): extract oauth model connections
